+++ app/api/domains/__init__.py
@@ -1,0 +1,5 @@
+from . import api
+from . import ava
+from . import cho
+from . import map
+from . import osu
