+++ app/api/domains/ava.py
@@ -1,0 +1,57 @@
+from pathlib import Path
+from typing import Literal
+from typing import Optional
+from typing import Union
+
+from fastapi import APIRouter
+from fastapi import Response
+from fastapi.responses import FileResponse
+
+import app.state
+import app.utils
+
+HTTPResponse = Optional[Union[bytes, tuple[int, bytes]]]
+
+""" ava: avatar server (for both ingame & external) """
+
+router = APIRouter(tags=["Avatars"])
+
+
+AVATARS_PATH = Path.cwd() / ".data/avatars"
+
+DEFAULT_AVATAR = AVATARS_PATH / "default.jpg"
+
+
+@router.get("/favicon.ico")
+async def get_favicon() -> Response:
+    return FileResponse(DEFAULT_AVATAR, media_type="image/ico")
+
+
+@router.get("/{user_id}.{extension}")
+async def get_avatar(
+    user_id: int,
+    extension: Literal["jpg", "jpeg", "png"],
+) -> Response:
+    avatar_path = AVATARS_PATH / f"{user_id}.{extension}"
+
+    if not avatar_path.exists():
+        avatar_path = DEFAULT_AVATAR
+
+    return FileResponse(
+        avatar_path,
+        media_type=app.utils.get_media_type(extension),
+    )
+
+
+@router.get("/{user_id}")
+async def get_avatar_osu(user_id: int) -> Response:
+    for extension in ("jpg", "jpeg", "png"):
+        avatar_path = AVATARS_PATH / f"{user_id}.{extension}"
+
+        if avatar_path.exists():
+            return FileResponse(
+                avatar_path,
+                media_type=app.utils.get_media_type(extension),
+            )
+
+    return FileResponse(DEFAULT_AVATAR, media_type="image/jpeg")
