+++ app/api/middlewares.py
@@ -1,0 +1,45 @@
+import time
+
+from cmyui.logging import Ansi
+from cmyui.logging import log
+from cmyui.logging import printc
+from cmyui.utils import magnitude_fmt_time
+from starlette.middleware.base import BaseHTTPMiddleware
+from starlette.middleware.base import RequestResponseEndpoint
+from starlette.requests import Request
+from starlette.responses import Response
+
+
+class MetricsMiddleware(BaseHTTPMiddleware):
+    async def dispatch(
+        self,
+        request: Request,
+        call_next: RequestResponseEndpoint,
+    ) -> Response:
+        start_time = time.perf_counter_ns()
+        response = await call_next(request)
+        end_time = time.perf_counter_ns()
+
+        time_elapsed = end_time - start_time
+
+        # TODO: add metric to datadog
+
+        col = (
+            Ansi.LGREEN
+            if 200 <= response.status_code < 300
+            else Ansi.LYELLOW
+            if 300 <= response.status_code < 400
+            else Ansi.LRED
+        )
+
+        url = f"{request.headers['host']}{request['path']}"
+
+        log(
+            f"[{request.method}] {response.status_code} {url}",
+            col,
+            end=" | ",
+        )
+        printc(f"Request took: {magnitude_fmt_time(time_elapsed)}", Ansi.LBLUE)
+
+        response.headers["process-time"] = str(round(time_elapsed) / 1e6)
+        return response
