+++ app/objects/beatmap.py
@@ -13,6 +13,7 @@ from typing import Optional
 from cmyui.logging import Ansi
 from cmyui.logging import log
 
+import app.settings
 import app.state
 import app.utils
 from app.constants.gamemodes import GameMode
@@ -34,10 +35,10 @@ IGNORED_BEATMAP_CHARS = dict.fromkeys(map(ord, r':\/*<>?"|'), None)
 
 async def osuapiv1_getbeatmaps(**params) -> Optional[list[dict[str, Any]]]:
     """Fetch data from the osu!api with a beatmap's md5."""
-    if app.state.settings.DEBUG:
+    if app.settings.DEBUG:
         log(f"Doing osu!api (getbeatmaps) request {params}", Ansi.LMAGENTA)
 
-    params["k"] = str(app.state.settings.OSU_API_KEY)
+    params["k"] = str(app.settings.OSU_API_KEY)
 
     async with app.state.services.http.get(OSUAPI_GET_BEATMAPS, params=params) as resp:
         if resp and resp.status == 200 and resp.content.total_bytes != 2:  # b'[]'
@@ -56,7 +57,7 @@ async def ensure_local_osu_file(
         or hashlib.md5(osu_file_path.read_bytes()).hexdigest() != bmap_md5
     ):
         # need to get the file from the osu!api
-        if app.state.settings.DEBUG:
+        if app.settings.DEBUG:
             log(f"Doing osu!api (.osu file) request {bmap_id}", Ansi.LMAGENTA)
 
         url = f"https://old.ppy.sh/osu/{bmap_id}"
@@ -65,7 +66,7 @@ async def ensure_local_osu_file(
                 if 400 <= resp.status < 500:
                     # client error, report this to cmyui
                     stacktrace = app.utils.get_appropriate_stacktrace()
-                    await app.utils.log_strange_occurrence(stacktrace)
+                    await app.state.services.log_strange_occurrence(stacktrace)
                 return False
 
             osu_file_path.write_bytes(await resp.read())
@@ -314,7 +315,7 @@ class Beatmap:
     @property
     def url(self) -> str:
         """The osu! beatmap url for `self`."""
-        return f"https://osu.{app.state.settings.DOMAIN}/beatmaps/{self.id}"
+        return f"https://osu.{app.settings.DOMAIN}/beatmaps/{self.id}"
 
     @property
     def embed(self) -> str:
@@ -600,7 +601,7 @@ class BeatmapSet:
     @property
     def url(self) -> str:  # same as above, just no beatmap id
         """The online url for this beatmap set."""
-        return f"https://osu.{app.state.settings.DOMAIN}/beatmapsets/{self.id}"
+        return f"https://osu.{app.settings.DOMAIN}/beatmapsets/{self.id}"
 
     @functools.cache
     def all_officially_ranked_or_approved(self) -> bool:
@@ -657,7 +658,7 @@ class BeatmapSet:
     async def _update_if_available(self) -> None:
         """Fetch newest data from the osu!api, check for differences
         and propogate any update into our cache & database."""
-        if not app.state.settings.OSU_API_KEY:
+        if not app.settings.OSU_API_KEY:
             return
 
         if api_data := await osuapiv1_getbeatmaps(s=self.id):
