+++ app/objects/models.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class OsuBeatmapRequestForm(BaseModel):
+    Filenames: list[str]
+    Ids: list[int]
