+++ app/objects/player.py
@@ -16,6 +16,7 @@ from cmyui.discord import Webhook
 from cmyui.logging import Ansi
 from cmyui.logging import log
 
+import app.settings
 import app.state
 import packets
 from app.constants.gamemodes import GameMode
@@ -35,7 +36,6 @@ from app.objects.menu import MenuFunction
 from app.objects.score import Grade
 from app.objects.score import Score
 from app.utils import escape_enum
-from app.utils import Geolocation
 from app.utils import pymysql_encode
 
 if TYPE_CHECKING:
@@ -255,7 +255,7 @@ class Player:
 
         self.achievements: set["Achievement"] = set()
 
-        self.geoloc: Geolocation = extras.get(
+        self.geoloc: app.state.services.Geolocation = extras.get(
             "geoloc",
             {
                 "latitude": 0.0,
@@ -321,7 +321,7 @@ class Player:
         # NOTE: this is currently never wiped because
         # domain & id cannot be changed in-game; if this
         # ever changes, it will need to be wiped.
-        return f"https://{app.state.settings.DOMAIN}/u/{self.id}"
+        return f"https://{app.settings.DOMAIN}/u/{self.id}"
 
     @cached_property
     def embed(self) -> str:
@@ -337,7 +337,7 @@ class Player:
         # NOTE: this is currently never wiped because
         # domain & id cannot be changed in-game; if this
         # ever changes, it will need to be wiped.
-        return f"https://a.{app.state.settings.DOMAIN}/{self.id}"
+        return f"https://a.{app.settings.DOMAIN}/{self.id}"
 
     @cached_property
     def full_name(self) -> str:
@@ -502,7 +502,7 @@ class Player:
 
         log(log_msg, Ansi.LRED)
 
-        if webhook_url := app.state.settings.DISCORD_AUDIT_LOG_WEBHOOK:
+        if webhook_url := app.settings.DISCORD_AUDIT_LOG_WEBHOOK:
             webhook = Webhook(webhook_url, content=log_msg)
             await webhook.post(app.state.services.http)
 
@@ -530,7 +530,7 @@ class Player:
 
         log(log_msg, Ansi.LRED)
 
-        if webhook_url := app.state.settings.DISCORD_AUDIT_LOG_WEBHOOK:
+        if webhook_url := app.settings.DISCORD_AUDIT_LOG_WEBHOOK:
             webhook = Webhook(webhook_url, content=log_msg)
             await webhook.post(app.state.services.http)
 
@@ -645,7 +645,7 @@ class Player:
     def leave_match(self) -> None:
         """Attempt to remove `self` from their match."""
         if not self.match:
-            if app.state.settings.DEBUG:
+            if app.settings.DEBUG:
                 log(f"{self} tried leaving a match they're not in?", Ansi.LYELLOW)
             return
 
@@ -749,7 +749,7 @@ class Player:
                 if c.can_read(p.priv):
                     p.enqueue(chan_info_packet)
 
-        if app.state.settings.DEBUG:
+        if app.settings.DEBUG:
             log(f"{self} joined {c}.")
 
         return True
@@ -780,7 +780,7 @@ class Player:
                 if c.can_read(p.priv):
                     p.enqueue(chan_info_packet)
 
-        if app.state.settings.DEBUG:
+        if app.settings.DEBUG:
             log(f"{self} left {c}.")
 
     def add_spectator(self, p: "Player") -> None:
