+++ app/state/services.py
@@ -1,10 +1,13 @@
 import asyncio
 import ipaddress
 import pickle
+import re
 import secrets
 from pathlib import Path
 from typing import AsyncGenerator
+from typing import AsyncIterator
 from typing import Optional
+from typing import TYPE_CHECKING
 from typing import TypedDict
 from typing import Union
 
@@ -12,29 +15,50 @@ import aiohttp
 import aioredis
 import cmyui
 import databases
+import datadog as datadog_module
 import datadog.threadstats.base as datadog_client
 import geoip2.database
+import orjson
+import pymysql
 from cmyui.logging import Ansi
 from cmyui.logging import log
 from cmyui.logging import printc
 from cmyui.logging import Rainbow
 
+import app.settings
 import app.state
 from app.constants.countries import country_codes
 
+if TYPE_CHECKING:
+    import databases.core
+
 IPAddress = Union[ipaddress.IPv4Address, ipaddress.IPv6Address]
+
 
 STRANGE_LOG_DIR = Path.cwd() / ".data/logs"
 GEOLOC_DB_FILE = Path.cwd() / "ext/GeoLite2-City.mmdb"
 
+VERSION_RGX = re.compile(r"^# v(?P<ver>\d+\.\d+\.\d+)$")
+SQL_UPDATES_FILE = Path.cwd() / "migrations/migrations.sql"
+
 
 """ session objects """
 
-http: "aiohttp.ClientSession"
-database: "databases.Database"
-redis: "aioredis.Redis"
-geoloc_db: Optional["geoip2.database.Reader"]
-datadog: Optional[datadog_client.ThreadStats]
+http = aiohttp.ClientSession(json_serialize=lambda x: orjson.dumps(x).decode())
+database = databases.Database(app.settings.DB_DSN)
+redis: aioredis.Redis = aioredis.from_url(app.settings.REDIS_DSN)
+
+geoloc_db: Optional["geoip2.database.Reader"] = None
+if GEOLOC_DB_FILE.exists():
+    geoloc_db = geoip2.database.Reader(GEOLOC_DB_FILE)
+
+datadog: Optional[datadog_client.ThreadStats] = None
+if app.settings.DATADOG_API_KEY and app.settings.DATADOG_APP_KEY:
+    datadog_module.initialize(
+        api_key=str(app.settings.DATADOG_API_KEY),
+        app_key=str(app.settings.DATADOG_APP_KEY),
+    )
+    datadog = datadog_client.ThreadStats()
 
 housekeeping_tasks: list[asyncio.Task] = []
 
@@ -52,8 +76,48 @@ class Geolocation(TypedDict):
     country: Country
 
 
+# fmt: off
+country_codes = {
+    "oc": 1,   "eu": 2,   "ad": 3,   "ae": 4,   "af": 5,   "ag": 6,   "ai": 7,   "al": 8,
+    "am": 9,   "an": 10,  "ao": 11,  "aq": 12,  "ar": 13,  "as": 14,  "at": 15,  "au": 16,
+    "aw": 17,  "az": 18,  "ba": 19,  "bb": 20,  "bd": 21,  "be": 22,  "bf": 23,  "bg": 24,
+    "bh": 25,  "bi": 26,  "bj": 27,  "bm": 28,  "bn": 29,  "bo": 30,  "br": 31,  "bs": 32,
+    "bt": 33,  "bv": 34,  "bw": 35,  "by": 36,  "bz": 37,  "ca": 38,  "cc": 39,  "cd": 40,
+    "cf": 41,  "cg": 42,  "ch": 43,  "ci": 44,  "ck": 45,  "cl": 46,  "cm": 47,  "cn": 48,
+    "co": 49,  "cr": 50,  "cu": 51,  "cv": 52,  "cx": 53,  "cy": 54,  "cz": 55,  "de": 56,
+    "dj": 57,  "dk": 58,  "dm": 59,  "do": 60,  "dz": 61,  "ec": 62,  "ee": 63,  "eg": 64,
+    "eh": 65,  "er": 66,  "es": 67,  "et": 68,  "fi": 69,  "fj": 70,  "fk": 71,  "fm": 72,
+    "fo": 73,  "fr": 74,  "fx": 75,  "ga": 76,  "gb": 77,  "gd": 78,  "ge": 79,  "gf": 80,
+    "gh": 81,  "gi": 82,  "gl": 83,  "gm": 84,  "gn": 85,  "gp": 86,  "gq": 87,  "gr": 88,
+    "gs": 89,  "gt": 90,  "gu": 91,  "gw": 92,  "gy": 93,  "hk": 94,  "hm": 95,  "hn": 96,
+    "hr": 97,  "ht": 98,  "hu": 99,  "id": 100, "ie": 101, "il": 102, "in": 103, "io": 104,
+    "iq": 105, "ir": 106, "is": 107, "it": 108, "jm": 109, "jo": 110, "jp": 111, "ke": 112,
+    "kg": 113, "kh": 114, "ki": 115, "km": 116, "kn": 117, "kp": 118, "kr": 119, "kw": 120,
+    "ky": 121, "kz": 122, "la": 123, "lb": 124, "lc": 125, "li": 126, "lk": 127, "lr": 128,
+    "ls": 129, "lt": 130, "lu": 131, "lv": 132, "ly": 133, "ma": 134, "mc": 135, "md": 136,
+    "mg": 137, "mh": 138, "mk": 139, "ml": 140, "mm": 141, "mn": 142, "mo": 143, "mp": 144,
+    "mq": 145, "mr": 146, "ms": 147, "mt": 148, "mu": 149, "mv": 150, "mw": 151, "mx": 152,
+    "my": 153, "mz": 154, "na": 155, "nc": 156, "ne": 157, "nf": 158, "ng": 159, "ni": 160,
+    "nl": 161, "no": 162, "np": 163, "nr": 164, "nu": 165, "nz": 166, "om": 167, "pa": 168,
+    "pe": 169, "pf": 170, "pg": 171, "ph": 172, "pk": 173, "pl": 174, "pm": 175, "pn": 176,
+    "pr": 177, "ps": 178, "pt": 179, "pw": 180, "py": 181, "qa": 182, "re": 183, "ro": 184,
+    "ru": 185, "rw": 186, "sa": 187, "sb": 188, "sc": 189, "sd": 190, "se": 191, "sg": 192,
+    "sh": 193, "si": 194, "sj": 195, "sk": 196, "sl": 197, "sm": 198, "sn": 199, "so": 200,
+    "sr": 201, "st": 202, "sv": 203, "sy": 204, "sz": 205, "tc": 206, "td": 207, "tf": 208,
+    "tg": 209, "th": 210, "tj": 211, "tk": 212, "tm": 213, "tn": 214, "to": 215, "tl": 216,
+    "tr": 217, "tt": 218, "tv": 219, "tw": 220, "tz": 221, "ua": 222, "ug": 223, "um": 224,
+    "us": 225, "uy": 226, "uz": 227, "va": 228, "vc": 229, "ve": 230, "vg": 231, "vi": 232,
+    "vn": 233, "vu": 234, "wf": 235, "ws": 236, "ye": 237, "yt": 238, "rs": 239, "za": 240,
+    "zm": 241, "me": 242, "zw": 243, "xx": 244, "a2": 245, "o1": 246, "ax": 247, "gg": 248,
+    "im": 249, "je": 250, "bl": 251, "mf": 252,
+}
+# fmt: on
+
+
 def fetch_geoloc_db(ip: IPAddress) -> Optional[Geolocation]:
     """Fetch geolocation data based on ip (using local db)."""
+    assert geoloc_db is not None
+
     res = geoloc_db.city(ip)
 
     if res.country.iso_code is not None:
@@ -106,13 +170,13 @@ async def log_strange_occurrence(obj: object) -> None:
     pickled_obj: bytes = pickle.dumps(obj)
     uploaded = False
 
-    if app.state.settings.AUTOMATICALLY_REPORT_PROBLEMS:
+    if app.settings.AUTOMATICALLY_REPORT_PROBLEMS:
         # automatically reporting problems to cmyui's server
         async with http.post(
             url="https://log.cmyui.xyz/",
             headers={
-                "Gulag-Version": app.state.settings.VERSION,
-                "Gulag-Domain": app.state.settings.DOMAIN,
+                "Gulag-Version": app.settings.VERSION,
+                "Gulag-Domain": app.settings.DOMAIN,
             },
             data=pickled_obj,
         ) as resp:
@@ -201,3 +265,103 @@ async def check_for_dependency_updates() -> None:
             "`python3.9 -m pip install -U <modules>`.",
             Ansi.LMAGENTA,
         )
+
+
+# sql migrations
+
+
+async def _get_current_sql_structure_version() -> Optional[cmyui.Version]:
+    """Get the last launched version of the server."""
+    res = await app.state.services.database.fetch_one(
+        "SELECT ver_major, ver_minor, ver_micro "
+        "FROM startups ORDER BY datetime DESC LIMIT 1",
+    )
+
+    if res:
+        return cmyui.Version(*map(int, res))
+
+
+async def run_sql_migrations() -> None:
+    """Update the sql structure, if it has changed."""
+    if not (current_ver := await _get_current_sql_structure_version()):
+        return  # already up to date (server has never run before)
+
+    latest_ver = cmyui.Version.from_str(app.settings.VERSION)
+
+    if latest_ver == current_ver:
+        return  # already up to date
+
+    # version changed; there may be sql changes.
+    content = SQL_UPDATES_FILE.read_text()
+
+    queries = []
+    q_lines = []
+
+    update_ver = None
+
+    for line in content.splitlines():
+        if not line:
+            continue
+
+        if line.startswith("#"):
+            # may be normal comment or new version
+            if r_match := VERSION_RGX.fullmatch(line):
+                update_ver = cmyui.Version.from_str(r_match["ver"])
+
+            continue
+        elif not update_ver:
+            continue
+
+        # we only need the updates between the
+        # previous and new version of the server.
+        if current_ver < update_ver <= latest_ver:
+            if line.endswith(";"):
+                if q_lines:
+                    q_lines.append(line)
+                    queries.append(" ".join(q_lines))
+                    q_lines = []
+                else:
+                    queries.append(line)
+            else:
+                q_lines.append(line)
+
+    if queries:
+        log(
+            f"Updating mysql structure (v{current_ver!r} -> v{latest_ver!r}).",
+            Ansi.LMAGENTA,
+        )
+
+    # XXX: so it turns out we can't use a transaction here (at least with mysql)
+    #      to roll back changes, as any structural changes to tables implicitly
+    #      commit: https://dev.mysql.com/doc/refman/5.7/en/implicit-commit.html
+    async with app.state.services.database.connection() as db_conn:
+        for query in queries:
+            try:
+                await db_conn.execute(query)
+            except pymysql.err.MySQLError as exc:
+                log(f"Failed: {query}", Ansi.GRAY)  # type: ignore
+                log(repr(exc))
+                log(
+                    "SQL failed to update - unless you've been "
+                    "modifying sql and know what caused this, "
+                    "please please contact cmyui#0425.",
+                    Ansi.LRED,
+                )
+                raise KeyboardInterrupt from exc
+        else:
+            # all queries executed successfully
+            await db_conn.execute(
+                "INSERT INTO startups (ver_major, ver_minor, ver_micro, datetime) "
+                "VALUES (:major, :minor, :micro, NOW())",
+                {
+                    "major": latest_ver.major,
+                    "minor": latest_ver.minor,
+                    "micro": latest_ver.micro,
+                },
+            )
+
+
+async def acquire_db_conn() -> AsyncIterator["databases.core.Connection"]:
+    """Decorator to acquire a database connection for a handler."""
+    async with database.connection() as conn:
+        yield conn
