+++ app/state/sessions.py
@@ -1,4 +1,8 @@
+import asyncio
 from typing import TYPE_CHECKING
+
+from cmyui.logging import Ansi
+from cmyui.logging import log
 
 from app.objects.collections import Channels
 from app.objects.collections import Clans
@@ -19,7 +23,35 @@ achievements: list["Achievement"] = []
 
 api_keys = {}
 
-ongoing_connections = set()
-housekeeping_tasks = set()
+housekeeping_tasks: set[asyncio.Task] = set()
 
 bot: "Player"
+
+
+# usecases
+
+
+async def cancel_housekeeping_tasks() -> None:
+    log(
+        f"-> Cancelling {len(housekeeping_tasks)} housekeeping tasks.",
+        Ansi.LMAGENTA,
+    )
+
+    # cancel housekeeping tasks
+    for task in housekeeping_tasks:
+        task.cancel()
+
+    await asyncio.gather(*housekeeping_tasks, return_exceptions=True)
+
+    loop = asyncio.get_running_loop()
+
+    for task in housekeeping_tasks:
+        if not task.cancelled():
+            if exception := task.exception():
+                loop.call_exception_handler(
+                    {
+                        "message": "unhandled exception during loop shutdown",
+                        "exception": exception,
+                        "task": task,
+                    },
+                )
