+++ bg_loops.py
@@ -4,6 +4,7 @@ import time
 from cmyui.logging import Ansi
 from cmyui.logging import log
 
+import app.settings
 import app.state
 import packets
 from app.constants.privileges import Privileges
@@ -32,7 +33,7 @@ async def initialize_housekeeping_tasks() -> None:
 async def _remove_expired_donation_privileges(interval: int) -> None:
     """Remove donation privileges from users with expired sessions."""
     while True:
-        if app.state.settings.DEBUG:
+        if app.settings.DEBUG:
             log("Removing expired donation privileges.", Ansi.LMAGENTA)
 
         expired_donors = await app.state.services.database.fetch_all(
