+++ scripts/start_server.sh
@@ -10,5 +10,7 @@ while ! nc -z localhost 6379; do
     sleep 0.25
 done
 
-SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
-python3.9 $SCRIPT_DIR/../main.py
+exec uvicorn --host 0.0.0.0 --port 1234 --no-access-log --reload app.api.init_api:asgi_app
+
+#SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+#python3.9 $SCRIPT_DIR/../main.py
